Add script-output visualization
The presented visualization shows the current status of the skill-map work.

The visualization itself uses `script-output.json` as the data set, this json
is being generated using a script that parses each readme file in the
for-now-called learning objectives project.

The profile data is being written in the `index.js` file as it was a late
addition to the main idea, but next versions should have these data as part of
the script output json.